fix: minor changes
