+++ packages/playground/src/components/filters/filters-bar/filter-variants/checkbox.tsx
@@ -1,4 +1,4 @@
-import { DropdownMenuCheckboxItem } from '@harnessio/canary'
+import { cn, DropdownMenuCheckboxItem, DropdownMenuItem, Icon, Input } from '@harnessio/canary'
 import { CheckboxFilterOption, FilterValue } from '../../types'
 import { UseFiltersReturn } from '../../use-filters'
 
@@ -6,28 +6,90 @@ interface CheckboxFilterProps {
   filter: FilterValue
   filterOption: CheckboxFilterOption
   onUpdateFilter: UseFiltersReturn['handleUpdateFilter']
+  searchQueries: UseFiltersReturn['searchQueries']
+  handleSearchChange: UseFiltersReturn['handleSearchChange']
 }
 
-const Checkbox = ({ filter, filterOption, onUpdateFilter }: CheckboxFilterProps) => {
+const Checkbox = ({ filter, filterOption, onUpdateFilter, searchQueries, handleSearchChange }: CheckboxFilterProps) => {
   return (
-    <div className="px-2 py-1">
-      {filterOption.options.map(option => (
-        <DropdownMenuCheckboxItem
-          className="pl-[34px]"
-          checked={filter.selectedValues.includes(option.value)}
-          onSelect={event => {
-            event?.preventDefault()
-            event?.stopPropagation()
-            const newValues = filter.selectedValues.includes(option.value)
-              ? filter.selectedValues.filter(v => v !== option.value)
-              : [...filter.selectedValues, option.value]
-            onUpdateFilter(filter.type, newValues)
-          }}
-          key={option.value}>
-          {option.label}
-        </DropdownMenuCheckboxItem>
-      ))}
-    </div>
+    <>
+      {filter.condition !== 'is_empty' && (
+        <div className="border-borders-1 border-b px-3 pb-2.5">
+          <div
+            className={cn(
+              'border-border-2 focus-within:border-borders-3 flex min-h-8 justify-between gap-x-1 rounded border px-2.5 py-[3px] outline-none transition-colors duration-200 focus-within:border',
+              {
+                'px-1': !!filter.selectedValues.length
+              }
+            )}>
+            <div className="flex flex-1 flex-wrap items-center gap-1">
+              {!!filter.selectedValues.length &&
+                filter.selectedValues.map(value => {
+                  const label = filterOption.options?.find(opt => opt.value === value)?.label
+                  return (
+                    <div className="bg-background-8 flex h-6 items-center gap-x-1.5 rounded px-2" key={value}>
+                      <span className="text-14 text-foreground-8">{label}</span>
+                      <button
+                        className="text-icons-1 hover:text-foreground-1 transition-colors duration-200"
+                        onClick={() => {
+                          const newValues = filter.selectedValues.filter(v => v !== value)
+                          onUpdateFilter(filter.type, newValues)
+                        }}>
+                        <Icon className="rotate-45" name="plus" size={10} />
+                      </button>
+                    </div>
+                  )
+                })}
+
+              <DropdownMenuItem className="p-0 focus:bg-transparent" asChild>
+                <Input
+                  className="h-6 flex-1 border-none outline-none hover:border-none focus:border-none focus-visible:ring-0"
+                  type="text"
+                  placeholder={filter.selectedValues.length > 0 ? '' : 'Select one or more options...'}
+                  value={searchQueries.filters[filter.type] || ''}
+                  onChange={e => handleSearchChange?.(filter.type, e.target.value, 'filters')}
+                  onClick={e => {
+                    e.preventDefault()
+                  }}
+                  onKeyDown={e => e.stopPropagation()}
+                />
+              </DropdownMenuItem>
+            </div>
+            {(!!filter.selectedValues.length || searchQueries.filters[filter.type]) && (
+              <button
+                className="text-foreground-4 hover:text-foreground-1 -mr-2.5 flex p-1.5 transition-colors duration-200"
+                onClick={() => {
+                  onUpdateFilter(filter.type, [])
+                  handleSearchChange?.(filter.type, '', 'filters')
+                }}>
+                <Icon className="rotate-45" name="plus" size={12} />
+              </button>
+            )}
+          </div>
+        </div>
+      )}
+
+      {!!filterOption.options.length && (
+        <div className="px-2 py-1">
+          {filterOption.options.map(option => (
+            <DropdownMenuCheckboxItem
+              className="pl-[34px]"
+              checked={filter.selectedValues.includes(option.value)}
+              onSelect={event => {
+                event?.preventDefault()
+                event?.stopPropagation()
+                const newValues = filter.selectedValues.includes(option.value)
+                  ? filter.selectedValues.filter(v => v !== option.value)
+                  : [...filter.selectedValues, option.value]
+                onUpdateFilter(filter.type, newValues)
+              }}
+              key={option.value}>
+              {option.label}
+            </DropdownMenuCheckboxItem>
+          ))}
+        </div>
+      )}
+    </>
   )
 }
